Use copy over os.Rename to avoid os restrictions
In some environments the temp directory used for downloads may not be
able to use os.Rename. In docker with caching enabled this fails, as the
directory is a volume mount, with the error: "invalid cross-device
link".

Another small fix is to hash the file using the io.Copy to avoid reading
the binary into memory.